RISP-Com/main_application/requirements.txt
aioice==0.7.6
aiortc==1.2.1
alabaster==0.7.12
altgraph==0.17.2
anyio==3.4.0
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
av==8.1.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachey==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dask==2022.1.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.2.0
docstring-parser==0.13
docutils==0.17.1
entrypoints==0.3
fonttools==4.28.5
freetype-py==2.2.0
fsspec==2022.1.0
future==0.18.2
google-crc32c==1.3.0
HeapDict==1.0.1
hsluv==5.0.2
idna==3.3
imageio==2.13.5
imagesize==1.3.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
intervaltree==3.1.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-packaging==0.11.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
locket==0.2.1
magicgui==0.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
napari==0.4.13
napari-console==0.0.4
napari-plugin-engine==0.2.0
napari-svg==0.1.6
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.6.3
notebook==6.4.6
npe2==0.1.1
numpy==1.21.5
numpydoc==1.1.0
open3d==0.14.1
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.0
Pint==0.18
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.9.0
psygnal==0.2.0
pycparser==2.21
pydantic==1.9.0
pyee==9.0.3
Pygments==2.11.1
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pylibsrtp==0.7.0
PyOpenGL==3.1.5
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytomlpp==1.0.10
pytz==2021.3
PyWavelets==1.2.0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
superqt==0.2.5.post1
terminado==0.12.1
testpath==0.5.0
tifffile==2021.11.2
tomlkit==0.8.0
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.7
vispy==0.9.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.13.3
zipp==3.7.0
